Use BigInteger instead of Integer for collection IDs.
We probably won't have more than sizeof(int32) collections, but it's
not inconceivable that we'd want to reserve chunks of address space in
the future, and then feel squeezed by int32.

This is also just good for consistency with other autoincrement IDs.

+++ python/lsst/daf/butler/registry/collections/synthIntKey.py
@@ -48,14 +48,14 @@ if TYPE_CHECKING:
 _TABLES_SPEC = CollectionTablesTuple(
     collection=ddl.TableSpec(
         fields=[
-            ddl.FieldSpec("collection_id", dtype=sqlalchemy.Integer, primaryKey=True, autoincrement=True),
+            ddl.FieldSpec("collection_id", dtype=sqlalchemy.BigInteger, primaryKey=True, autoincrement=True),
             ddl.FieldSpec("name", dtype=sqlalchemy.String, length=64, nullable=False),
             ddl.FieldSpec("type", dtype=sqlalchemy.SmallInteger, nullable=False),
         ],
         unique=[("name",)],
     ),
-    run=makeRunTableSpec("collection_id", sqlalchemy.Integer),
-    collection_chain=makeCollectionChainTableSpec("collection_id", sqlalchemy.Integer),
+    run=makeRunTableSpec("collection_id", sqlalchemy.BigInteger),
+    collection_chain=makeCollectionChainTableSpec("collection_id", sqlalchemy.BigInteger),
 )
